move yearly table generator to processing rather than uploads
Single responsibility principle
