+++ pipeline/climatology_uploads.py
@@ -24,37 +24,8 @@ logging.basicConfig(filename='uploads_logger.log', encoding='utf-8', level=loggi
 @dataclass
 class ClimatologyUploads(Climatology):
     climatology_url: str
+    schema: str = "climatology"
     
-    def climatology_yearly_table_generator(
-        self,
-    ):
-        self._climatology_pathways(self.climatology_url)
-        """Aggregates monthly climatology predictions into a yearly table.
-        Processing steps: 
-        * Add month int month column (1-12)
-        * Sort values by administrative identifier and month
-        """
-
-        if not os.path.exists(f"{self.time_series}/{self.climatology}_yearly.csv"):
-            zs_files = glob.glob(os.path.join(self.zonal_statistics, '*.csv'))
-            
-            li = []
-            logger.info(f"Creating a yearly dataset for {self.climatology}")
-
-            for file in zs_files:
-                with open(f"{file}", 'r') as f:
-                    month = re.search('_\d{1,2}', file).group(0)
-                    month = month.replace("_", "")
-                    df = pd.read_csv(f, index_col=None, header=0)
-                    df['month'] = int(month)
-                    li.append(df)
-
-                data = pd.concat(li, axis=0, ignore_index=True)
-                data.sort_values(by=["OBJECTID_1", "month"], inplace=True)
-                data.to_csv(f"{self.time_series}/{self.climatology}_yearly.csv", index=False)
-                
-        else:
-            logger.info(f"Yearly time appended dataset exists for {self.climatology}")
     def upload_to_db(
         self,
         engine: Engine
@@ -68,6 +39,7 @@ class ClimatologyUploads(Climatology):
         Args:
             engine (Engine): Engine connection to postgres db
         """
+        self._climatology_pathways(self.climatology_url)
 
         table = self.climatology.lower()
 
@@ -109,7 +81,6 @@ def local_to_postgres_flow(
 
     for url in climatologies:
         cmip_temp = ClimatologyUploads(climatology_url=url)
-        cmip_temp.climatology_yearly_table_generator()
         cmip_temp.upload_to_db(engine=engine)
 
 if __name__ == "__main__":
